+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/Student.java
@@ -26,7 +26,6 @@ public class Student {
     @JsonBackReference
     private Course course;
 
-    //TODO : Falta lista de Courses
 
     //TODO : Falta lista de Parents y Auth